tests/wpt/web-platform-tests/app-history/currentchange-event/currentchange-navigate-from-initial-about-blank-same-doc-popup.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  const w = window.open("about:blank#1");
  w.appHistory.oncurrentchange = t.unreached_func("currentchange should not fire");

  assert_equals(w.location.href, "about:blank#1");
  w.location.href = "about:blank#2";
  assert_equals(w.location.href, "about:blank#2");

  await new Promise(resolve => t.step_timeout(resolve, 10));
}, "AppHistoryCurrentChangeEvent does not fire when navigating away from the initial about:blank (popup window)");
</script>
